ISSUE #91 * Improve tests.
